refactor: MVC 디렉토리 구조로 변경
- MVC에 맞게 디렉토리 구조 변경
