+++ src/app.js
@@ -4,9 +4,8 @@ const logger = require('morgan');
 const cookieParser = require('cookie-parser');
 const createError = require('http-errors');
 const session = require('./middlewares/session');
-const indexRouter = require('./routes/index');
-const apiRouter = require('./routes/api');
-const signUpRouter = require('./routes/sign-up');
+const router = require('./routes');
+
 const app = express();
 
 app.set('views', path.join(__dirname, 'views'));
@@ -19,9 +18,7 @@ app.use(cookieParser());
 app.use(express.static(path.join(__dirname, 'public')));
 
 app.use(session);
-app.use('/', indexRouter);
-app.use('/api', apiRouter);
-app.use('/sign-up', signUpRouter);
+app.use('/', router);
 
 app.use((req, res, next) => {
   next(createError(404));
@@ -32,7 +29,7 @@ app.use((err, req, res) => {
   res.locals.error = req.app.get('env') === 'development' ? err : {};
 
   res.status(err.status || 500);
-  res.render('error');
+  res.render('404');
 });
 
 module.exports = app;
