+++ src/routes/auth/sign-up.js
@@ -1,16 +1,8 @@
 const express = require('express');
 const router = express.Router();
 
-router.use((req, res, next) => {
-  const { session } = req;
-  if (session) {
-    res.redirect('/');
-  }
-  next();
-});
-
 router.get('/', (req, res) => {
-  res.redirect('/sign-up/agree');
+  res.redirect('/auth/sign-up/agree');
 });
 
 router.get('/agree', (req, res) => {
